Completed refactor
