+++ core_api/src/format.py
@@ -3,8 +3,8 @@ from uuid import UUID
 
 from langchain_core.documents.base import Document
 
-from redbox.models.file import Chunk, Metadata, combine_documents
-from redbox.storage import ElasticsearchStorageHandler
+from redbox.transform import combine_documents
+from redbox.models.file import Chunk, encoding
 
 
 def format_chunks(chunks: list[Chunk]) -> str:
@@ -34,21 +34,11 @@ def reduce_chunks_by_tokens(chunks: list[Document] | None, chunk: Document, max_
         return [chunk]
 
     last_chunk = chunks[-1]
-    if chunk.metadata['token_count'] + last_chunk.metadata['token_count'] <= max_tokens:
+    # Backwards compatible with worker which didn't store token_count
+    chunk_tokens = chunk.metadata["_source"]["metadata"].get('token_count') or len(encoding.encode(chunk.page_content))
+    last_chunk_tokens = last_chunk.metadata["_source"]["metadata"].get('token_count') or len(encoding.encode(last_chunk.page_content))
+    if chunk_tokens + last_chunk_tokens <= max_tokens:
         chunks[-1] = combine_documents(last_chunk, chunk)
     else:
         chunks.append(chunk)
     return chunks
-
-
-def get_file_chunked_to_tokens(
-    file_uuid: UUID, user_uuid: UUID, storage_handler: ElasticsearchStorageHandler, max_tokens: int | None = None
-) -> list[Chunk]:
-    """Gets a file as larger document-sized Chunks, splitting it by max_tokens."""
-    n = max_tokens or float("inf")
-    chunks_unsorted = storage_handler.get_file_chunks(parent_file_uuid=file_uuid, user_uuid=user_uuid)
-    chunks_sorted = sorted(chunks_unsorted, key=lambda x: x.index)
-
-    reduce_chunk_n = partial(reduce_chunks_by_tokens, max_tokens=n)
-
-    return reduce(lambda cs, c: reduce_chunk_n(cs, c), chunks_sorted, [])
