+++ redbox/models/file.py
@@ -147,19 +147,3 @@ class FileStatus(BaseModel):
     chunk_statuses: list[ChunkStatus] | None
 
 
-def combine_documents(a: Document, b: Document):
-    def combine_values(field_name):
-        return a.metadata[field_name] + b.metadata[field_name]
-    
-    combined_content = a.page_content + b.page_content
-    combined_metadata = a.metadata.copy()
-    combined_metadata["token_count"] = len(encoding.encode(combined_content.text))
-    combined_metadata['page_number'] = combine_values("page_number")
-    combined_metadata['languages'] = combine_values("languages")
-    combined_metadata['link_texts'] = combine_values("link_texts")
-    combined_metadata['link_urls'] = combine_values("link_urls")
-    combined_metadata['links'] = combine_values("links")
-    yield Document(
-        page_content=combined_content,
-        metadata=combined_metadata
-    )
